+++ packages/lyric/src/myplayer-app.ts
@@ -1,4 +1,5 @@
 import { TTMLParser, toAmllLyrics } from "@applemusic-like-lyrics/ttml";
+import { stringifyLrcA2 } from "./formats/lrca2";
 import { stringifyTTML } from "./formats/ttml";
 import type { LyricLine, LyricWord, TTMLLyric } from "./types";
 
@@ -7,12 +8,18 @@ export { decryptQrcHex, encryptQrcHex } from "./formats/eqrc";
 export { parseEslrc, stringifyEslrc } from "./formats/eslrc";
 export { parseLqe, stringifyLqe } from "./formats/lqe";
 export { parseLrc, stringifyLrc } from "./formats/lrc";
-export { parseLrcA2, stringifylrcA2 } from "./formats/lrca2";
+export { parseLrcA2, stringifyLrcA2 } from "./formats/lrca2";
 export { parseLyl, stringifyLyl } from "./formats/lyl";
 export { parseLys, stringifyLys } from "./formats/lys";
 export { parseQrc, stringifyQrc } from "./formats/qrc";
 export { parseYrc, stringifyYrc } from "./formats/yrc";
 export type { LyricLine, LyricWord, TTMLLyric } from "./types";
+
+export function stringifylrcA2(
+	...args: Parameters<typeof stringifyLrcA2>
+): ReturnType<typeof stringifyLrcA2> {
+	return stringifyLrcA2(...args);
+}
 
 const TTML_NS = "http://www.w3.org/ns/ttml";
 const ROLE_ATTR_NAMES = ["role", "ttm:role"];