+++ packages/playground/core/src/components/SidebarShell.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { PackageIcon } from "lucide-vue-next";
+import amllLogoSvg from "@/assets/amll-logo.svg";
 import {
 	Sidebar,
 	SidebarContent,
@@ -21,9 +21,9 @@ const coreVersion = __AMLL_CORE_VERSION__;
 		<SidebarHeader class="border-b border-sidebar-border p-3">
 			<div class="flex gap-2 items-center">
 				<div
-					class="flex aspect-square size-8 items-center justify-center rounded-lg bg-primary text-sidebar-primary-foreground"
+					class="flex aspect-square size-8 items-center justify-center rounded-md bg-primary text-sidebar-primary-foreground overflow-hidden"
 				>
-					<PackageIcon class="size-4" />
+					<img :src="amllLogoSvg">
 				</div>
 				<div class="flex flex-col gap-1 leading-none">
 					<span class="font-medium"> AMLL Playground </span>